Changed loot rate of mirrors
Still not seeing them in drops but I the logic looks like its working. Best guess is the drop rate is super low (maybe 1%). Seems the drop rate is worst the more mods that add loot to the tables. Configs might midigate this for well made packs. Though a better solution is need in the long term. Recipes would be the best option. As this entires the item can be accessed.